pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "dnsctl-app"
version = "1.2.0"
authors = [{ name = "Dhivijit Koppuravuri", email = "me@dhivijit.dev" }]
description = "Secure, version-controlled DNS management tool with CLI and GUI"
readme = "README.md"
license = "MIT"
requires-python = ">=3.11"
keywords = ["dns", "devops", "infrastructure", "git", "cli", "security"]
classifiers = [
    "Development Status :: 5 - Production/Stable",
    "Environment :: Console",
    "Environment :: X11 Applications :: Qt",
    "Intended Audience :: Developers",
    "Intended Audience :: System Administrators",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Internet :: Name Service (DNS)",
    "Topic :: System :: Systems Administration",
    "Topic :: Utilities",
]
dependencies = [
    "requests>=2.31",
    "cryptography>=41.0",
    "keyring>=24.0",
    "click>=8.1",
    "GitPython>=3.1",
    "textual>=0.60",
]

[project.optional-dependencies]
gui = [
    "PyQt6>=6.6",
    "qt-material>=2.14",
    "qtawesome>=1.3",
]
dev = [
    "pytest>=7.4",
    "pyinstaller>=6.0",
]

[project.urls]
"Source" = "https://github.com/dhivijit/dnsctl"
"Issue Tracker" = "https://github.com/dhivijit/dnsctl/issues"

[project.scripts]
dnsctl = "dnsctl.cli.main:main"

[project.gui-scripts]
dnsctl-g = "dnsctl.gui.app:main"

[tool.setuptools.packages.find]
where = ["."]
include = ["dnsctl*"]

[tool.setuptools.package-data]
dnsctl = ["icon.png"]
"dnsctl.gui" = ["ui/*.ui"]

[tool.pytest.ini_options]
testpaths = ["tests"]
